1. Used index uk_device_template_key_tenant_id instead 2. Removed templateType 3. Removed some unused codes
