+++ core/context/src/main/java/com/milesight/beaveriot/context/constants/IntegrationConstants.java
@@ -28,6 +28,11 @@ public class IntegrationConstants {
     public static final String INTEGRATION_DEVICE_KEY_FORMAT = "{0}.device.{1}";
 
     /**
+     * Integration key format， format: {integrationName}.device-template.{deviceTemplateIdentifier}
+     */
+    public static final String INTEGRATION_DEVICE_TEMPLATE_KEY_FORMAT = "{0}.device-template.{1}";
+
+    /**
      * Integration entity key format， format: {integrationName}.device.{deviceIdentifier}.{entityIdentifier}
      */
     public static final String INTEGRATION_DEVICE_IDENTIFY_ENTITY_KEY_FORMAT = "{0}.device.{1}.{2}";
@@ -43,6 +48,12 @@ public class IntegrationConstants {
         Assert.notNull(integrationId, "integrationId must not be null");
         Assert.notNull(deviceIdentifier, "deviceIdentifier must not be null");
         return MessageFormat.format(INTEGRATION_DEVICE_KEY_FORMAT, integrationId, deviceIdentifier);
+    }
+
+    public static String formatIntegrationDeviceTemplateKey(String integrationId, String deviceTemplateIdentifier) {
+        Assert.notNull(integrationId, "integrationId must not be null");
+        Assert.notNull(deviceTemplateIdentifier, "deviceTemplateIdentifier must not be null");
+        return MessageFormat.format(INTEGRATION_DEVICE_TEMPLATE_KEY_FORMAT, integrationId, deviceTemplateIdentifier);
     }
 
     public static String formatIntegrationDeviceEntityKey(String integrationId, String deviceIdentifier, String entityIdentifier) {
