+++ core/context/src/main/java/com/milesight/beaveriot/context/integration/model/config/IntegrationConfig.java
@@ -1,6 +1,9 @@
 package com.milesight.beaveriot.context.integration.model.config;
 
-import com.milesight.beaveriot.context.integration.model.*;
+import com.milesight.beaveriot.context.integration.model.Device;
+import com.milesight.beaveriot.context.integration.model.DeviceBuilder;
+import com.milesight.beaveriot.context.integration.model.Integration;
+import com.milesight.beaveriot.context.integration.model.IntegrationBuilder;
 import lombok.Data;
 
 import java.util.ArrayList;
@@ -26,12 +29,6 @@ public class IntegrationConfig {
 
     private String entityIdentifierDeleteDevice;
 
-    private String entityIdentifierAddDeviceTemplate;
-
-    private String entityIdentifierUpdateDeviceTemplate;
-
-    private String entityIdentifierDeleteDeviceTemplate;
-
     private List<DeviceConfig> initialDevices = new ArrayList<>();
 
     private List<EntityConfig> initialEntities = new ArrayList<>();
@@ -56,9 +53,6 @@ public class IntegrationConfig {
                 .enabled(enabled)
                 .entityIdentifierAddDevice(entityIdentifierAddDevice)
                 .entityIdentifierDeleteDevice(entityIdentifierDeleteDevice)
-                .entityIdentifierAddDeviceTemplate(entityIdentifierAddDeviceTemplate)
-                .entityIdentifierUpdateDeviceTemplate(entityIdentifierUpdateDeviceTemplate)
-                .entityIdentifierDeleteDeviceTemplate(entityIdentifierDeleteDeviceTemplate)
                 .end()
                 .initialEntities(()-> initialEntities.stream().map(entityConfig -> entityConfig.toEntity()).toList())
                 .initialDevices(devices)
