--- core/context/src/main/java/com/milesight/beaveriot/context/model/DeviceTemplateType.java
@@ -1,6 +1,0 @@
-package com.milesight.beaveriot.context.model;
-
-public enum DeviceTemplateType {
-    JSON,
-    BYTE
-}
