+++ services/device-template/device-template-service/src/main/resources/template/default_device_template.yaml
@@ -1,4 +1,3 @@
-template_type: JSON
 definition:
   input:
     type: object
